Cleaned up project structure and linked the UI to SetCubeBuilder
The filenames were not very clear for how the project was named. The UI
Renamed ContactEditorUI.java to SetCubeBuilderUI.java to clear this up
and to make them more consistent. The main method was moved into
SetCubeBuilderUI.java, so there was no need for a driver file any more.
The UI will now call SetCubeBuilder after the submit button is pressed.
This will generate the correct file. Overall it seems that the vast
majority of the functionality is working.