Added a second scene and cleaned up dialog box
I went back in NetBeans and built a second scene using card layout. The
second scene makes it clear that the program has finished downloading
data and producing a card list. Additionally the second scene adds a
close button and text that reminds the user what the filename was.
Overall I think this is now ready for public use.